restructure code a bit and make icons smaller
